Factor out ScrollZone component.
